# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "KnowledgeCitationResponse": {
        "properties": {
          "excerpt": {
            "title": "Excerpt",
            "type": "string"
          },
          "heading": {
            "title": "Heading",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "source",
          "heading",
          "score",
          "excerpt",
          "version"
        ],
        "title": "KnowledgeCitationResponse",
        "type": "object"
      },
      "KnowledgeDocumentCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "maxLength": 524288,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "expected_revision": {
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "source": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "expected_revision",
          "source",
          "content"
        ],
        "title": "KnowledgeDocumentCreateRequest",
        "type": "object"
      },
      "KnowledgeDocumentListResponse": {
        "properties": {
          "enabled_count": {
            "title": "Enabled Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeDocumentSummaryResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "total_enabled_bytes": {
            "title": "Total Enabled Bytes",
            "type": "integer"
          }
        },
        "required": [
          "revision",
          "total",
          "enabled_count",
          "total_enabled_bytes",
          "items"
        ],
        "title": "KnowledgeDocumentListResponse",
        "type": "object"
      },
      "KnowledgeDocumentResponse": {
        "properties": {
          "archived": {
            "title": "Archived",
            "type": "boolean"
          },
          "byte_size": {
            "title": "Byte Size",
            "type": "integer"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "content_sha256": {
            "title": "Content Sha256",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "current_version": {
            "title": "Current Version",
            "type": "integer"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "updated_by": {
            "title": "Updated By",
            "type": "string"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeVersionResponse"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "id",
          "source",
          "title",
          "enabled",
          "archived",
          "current_version",
          "content_sha256",
          "byte_size",
          "created_by",
          "updated_by",
          "created_at",
          "updated_at",
          "content",
          "versions"
        ],
        "title": "KnowledgeDocumentResponse",
        "type": "object"
      },
      "KnowledgeDocumentStateRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_document_version": {
            "minimum": 1.0,
            "title": "Expected Document Version",
            "type": "integer"
          },
          "expected_revision": {
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          }
        },
        "required": [
          "expected_revision",
          "expected_document_version"
        ],
        "title": "KnowledgeDocumentStateRequest",
        "type": "object"
      },
      "KnowledgeDocumentSummaryResponse": {
        "properties": {
          "archived": {
            "title": "Archived",
            "type": "boolean"
          },
          "byte_size": {
            "title": "Byte Size",
            "type": "integer"
          },
          "content_sha256": {
            "title": "Content Sha256",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "current_version": {
            "title": "Current Version",
            "type": "integer"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "updated_by": {
            "title": "Updated By",
            "type": "string"
          }
        },
        "required": [
          "id",
          "source",
          "title",
          "enabled",
          "archived",
          "current_version",
          "content_sha256",
          "byte_size",
          "created_by",
          "updated_by",
          "created_at",
          "updated_at"
        ],
        "title": "KnowledgeDocumentSummaryResponse",
        "type": "object"
      },
      "KnowledgeDocumentUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "maxLength": 524288,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "expected_document_version": {
            "minimum": 1.0,
            "title": "Expected Document Version",
            "type": "integer"
          },
          "expected_revision": {
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "source": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "expected_revision",
          "source",
          "content",
          "expected_document_version"
        ],
        "title": "KnowledgeDocumentUpdateRequest",
        "type": "object"
      },
      "KnowledgeMutationResponse": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/KnowledgeDocumentResponse"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "revision",
          "document"
        ],
        "title": "KnowledgeMutationResponse",
        "type": "object"
      },
      "KnowledgeSearchResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeCitationResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "query": {
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query",
          "items"
        ],
        "title": "KnowledgeSearchResponse",
        "type": "object"
      },
      "KnowledgeVersionResponse": {
        "properties": {
          "byte_size": {
            "title": "Byte Size",
            "type": "integer"
          },
          "content_sha256": {
            "title": "Content Sha256",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version",
          "content_sha256",
          "byte_size",
          "created_by",
          "created_at"
        ],
        "title": "KnowledgeVersionResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OpenReviewer API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/knowledge/documents": {
      "get": {
        "operationId": "list_knowledge_documents_api_v1_knowledge_documents_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 128,
              "maximum": 128,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeDocumentListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Knowledge Documents"
      },
      "post": {
        "operationId": "create_knowledge_document_api_v1_knowledge_documents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeDocumentCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Knowledge Document"
      }
    },
    "/api/v1/knowledge/documents/{document_id}": {
      "get": {
        "operationId": "get_knowledge_document_api_v1_knowledge_documents__document_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeDocumentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Knowledge Document"
      },
      "put": {
        "operationId": "update_knowledge_document_api_v1_knowledge_documents__document_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeDocumentUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Knowledge Document"
      }
    },
    "/api/v1/knowledge/documents/{document_id}/archive": {
      "post": {
        "operationId": "archive_knowledge_document_api_v1_knowledge_documents__document_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeDocumentStateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Knowledge Document"
      }
    },
    "/api/v1/knowledge/documents/{document_id}/restore": {
      "post": {
        "operationId": "restore_knowledge_document_api_v1_knowledge_documents__document_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeDocumentStateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Knowledge Document"
      }
    },
    "/api/v1/knowledge/documents/{document_id}/versions/{version}/restore": {
      "post": {
        "operationId": "restore_knowledge_document_version_api_v1_knowledge_documents__document_id__versions__version__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Version",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeDocumentStateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Knowledge Document Version"
      }
    },
    "/api/v1/knowledge/search": {
      "get": {
        "operationId": "search_knowledge_api_v1_knowledge_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 500,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Knowledge"
      }
    },
...
  }
}